Move new cloner to DarcLib
